refactor: use transactional migrations now that upstream issue is fixed

+++ core/migrations/mysql/Version20250807232311.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace DoctrineMigrations;
 
 use Doctrine\DBAL\Schema\Schema;
-use Doctrine\Migrations\AbstractMigration;
 
 final class Version20250807232311 extends AbstractMigration
 {
@@ -16,18 +15,17 @@ final class Version20250807232311 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $this->addSql('CREATE TABLE api_token (id INT AUTO_INCREMENT NOT NULL, name VARCHAR(255) NOT NULL, created_at DATETIME NOT NULL, expires_at DATETIME NULL, token_hash VARCHAR(255) NOT NULL, last_used DATETIME DEFAULT NULL, user_id INT NOT NULL, INDEX IDX_7BA2F5EBA76ED395 (user_id), PRIMARY KEY (id)) DEFAULT CHARACTER SET utf8mb4');
-        $this->addSql('ALTER TABLE api_token ADD CONSTRAINT FK_7BA2F5EBA76ED395 FOREIGN KEY (user_id) REFERENCES salt_user (id)');
+        if (!$schema->hasTable('api_token')) {
+            $this->addSql('CREATE TABLE api_token (id INT AUTO_INCREMENT NOT NULL, name VARCHAR(255) NOT NULL, created_at DATETIME NOT NULL, expires_at DATETIME NULL, token_hash VARCHAR(255) NOT NULL, last_used DATETIME DEFAULT NULL, user_id INT NOT NULL, INDEX IDX_7BA2F5EBA76ED395 (user_id), PRIMARY KEY (id)) DEFAULT CHARACTER SET utf8mb4');
+            $this->addSql('ALTER TABLE api_token ADD CONSTRAINT FK_7BA2F5EBA76ED395 FOREIGN KEY (user_id) REFERENCES salt_user (id)');
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE api_token DROP FOREIGN KEY FK_7BA2F5EBA76ED395');
-        $this->addSql('DROP TABLE api_token');
-    }
-
-    public function isTransactional(): bool
-    {
-        return false;
+        if ($schema->hasTable('api_token')) {
+            $this->addSql('ALTER TABLE api_token DROP FOREIGN KEY FK_7BA2F5EBA76ED395');
+            $this->addSql('DROP TABLE api_token');
+        }
     }
 }
